address PR #29 review
- Split `src/install/resolve.ts` (372 lines) into a directory:
  `resolve/index.ts` (160 lines, orchestration) and
  `resolve/enqueue.ts` (239 lines, per-ref helpers). Both under the
  file-size cap or close to it; the orchestrator no longer drags in
  every leaf dependency.
- Fix stale doc example: README and site Taps.tsx said
  `acme/marketing/copy-review`, but the example tree only contains
  `email-outreach` and `social-posts`. Swapped to `email-outreach`.
- Drop the re-export from `src/install/resolve-ref/index.ts` and
  import `formatCandidate` / `shortLabelFor` directly from
  `./format.ts` in `ambiguity-prompt.ts`. CLAUDE.md's re-export
  guidance is to avoid them as architectural layers.
- Comment `if (ctx.flags.yes) return raw` in the ambiguity prompt
  to explain why `--yes` doesn't pick a default here (unlike the
  tap/skill collision path).
- PRD §8.3 and C-NS-06/07/08 rewritten to reflect that `--tap`,
  `--bundle`, `--skill` are presence flags — the name comes from
  the positional.
